https://chronobank.atlassian.net/browse/MINT-241 - middleware - update event listener
