+++ index.js
@@ -1,7 +1,8 @@
 const mongoose = require('mongoose'),
   config = require('./config.json'),
   blockModel = require('./models').blockModel,
-  contractsCtrl = require('./controllers').contractsCtrl(),
+  contractsCtrl = require('./controllers').contractsCtrl,
+  eventsCtrl = require('./controllers').eventsCtrl,
   emitter = require('events'),
   _ = require('lodash'),
   plugins = require('./plugins'),
@@ -16,37 +17,51 @@ const mongoose = require('mongoose'),
 
 mongoose.connect(config.mongo.uri);
 
-Promise.all([
-  contractsCtrl.initEmitter,
-  contractsCtrl.contracts,
-  blockModel.findOne().sort('-block')
-]).spread((components, contracts, block) => {
+contractsCtrl()
+  .then(data =>
+    Promise.all([
+      {
+        eventModels: eventsCtrl(data.instances).eventModels,
+        contracts: data,
+      },
+      blockModel.findOne().sort('-block')
+    ])
+  )
+  .then((data) => {
 
-  block = _.chain(block).get('block', 0).add(0).value();
-  let eventEmitter = new emitter();
+    let contracts = data[0].contracts.contracts;
+    let contract_instances = data[0].contracts.instances;
+    let eventModels = data[0].eventModels;
+    let block = data[1];
 
-  /** initialize eventEmitter and watcher for eth contracts **/
-  _.chain(components).values()
-    .forEach(instance => {
-      let events = instance.allEvents({fromBlock: block, toBlock: 'latest'});
+    block = _.chain(block).get('block', 0).add(0).value();
+    let addr = contract_instances.MultiEventsHistory.address;
+    let eventEmitter = new emitter();
 
-      events.watch((error, result) => {
-        if (!_.has(result, 'event') || !contractsCtrl.eventModels[result.event] || result.blockNumber <= block) {
-          return;
-        }
-        let new_event = new contractsCtrl.eventModels[result.event](result.args);
-        new_event.save();
-        let new_block = new blockModel({block: result.blockNumber});
-        new_block.save();
-        eventEmitter.emit(result.event, result.args);
-      });
+    _.chain(contracts).values()
+      .forEach(instance => {
+        let events = instance.at(addr).allEvents({fromBlock: block, toBlock: 'latest'});
 
-    })
-    .value();
+        events.watch((error, result) => {
 
-  /** register all plugins **/
-  _.chain(plugins).values()
-    .forEach(plugin=> plugin(eventEmitter, contracts, contractsCtrl.eventModels))
-    .value();
+          if (!_.has(result, 'event') || !eventModels[result.event] || result.blockNumber <= block) {
+            return;
+          }
+          let new_event = new eventModels[result.event](result.args);
+          new_event.save();
+          let new_block = new blockModel({block: result.blockNumber});
+          new_block.save();
+          eventEmitter.emit(result.event, result.args);
+        });
 
-});
+      })
+      .value();
+
+    // register all plugins
+/*
+    _.chain(plugins).values()
+      .forEach(plugin => plugin(eventEmitter, contracts, eventModels))
+      .value();
+*/
+
+  });
