+++ tests/core/eventsCore.test.js
@@ -7,6 +7,7 @@ const config = require('../../config.json'),
   Promise = require('bluebird'),
   provider = new Web3.providers.HttpProvider(config.web3.url),
   helpers = require('../helpers'),
+  contractsCtrl = require('../../controllers').contractsCtrl,
   chronoBankPlatformEmitter_definition = require("../../SmartContracts/build/contracts/ChronoBankPlatformEmitter"),
   chronoBankPlatform_definition = require("../../SmartContracts/build/contracts/ChronoBankPlatform"),
   chronomint_definition = require("../../SmartContracts/build/contracts/LOCManager"),
@@ -29,11 +30,12 @@ beforeAll(() => {
       c.setProvider(provider);
     });
 
-  return helpers.setup()
+  return contractsCtrl()
     .then((data) => {
       _.merge(contracts_instances, data.instances);
+      _.merge(contracts, data.contracts);
       return Promise.all([
-        //chronoBankPlatformEmitter.at(helpers.setup.eventsHistory.address),
+        chronoBankPlatformEmitter.at(contracts_instances.EventsHistory.address),
         chronoMint.at(contracts_instances.MultiEventsHistory.address)
       ])
     })
@@ -90,7 +92,8 @@ test('add new loc', () => {
 
 test('fetch changes for loc via getLoc', () =>
   new Promise(res => {
-    contracts.mint.allEvents({fromBlock: 0}).watch((err, result) => {
+    contracts.LOCManager.at(contracts_instances.MultiEventsHistory.address)
+      .allEvents({fromBlock: 0}).watch((err, result) => {
       if (result && result.event === 'NewLOC') {
         expect(result.args.locName).toBeDefined();
         contracts_instances.LOCManager.getLOCByName(result.args.locName)
@@ -103,98 +106,82 @@ test('fetch changes for loc via getLoc', () =>
     });
   })
 );
-/*
 
- test('validate hash in mongo', () =>
- Promise.delay(2000)
- .then(() =>
- mongoose.model('NewLOC', new mongoose.Schema({
- locName: {type: mongoose.Schema.Types.Mixed}
- }
- )).findOne({locName: factory.Loc.name})
- )
- .then(result => {
- console.log(result);
- console.log(factory.Loc)
- expect(result).toBeDefined();
- expect(result.locName).toBeDefined();
- return Promise.resolve();
- })
- );
- */
-/*
- test('update loc', () => {
+test('validate hash in mongo', () =>
+  Promise.delay(2000)
+    .then(() =>
+      mongoose.model('NewLOC', new mongoose.Schema({
+          locName: {type: mongoose.Schema.Types.Mixed}
+        }
+      )).findOne({locName: factory.Loc.name})
+    )
+    .then(result => {
+      expect(result).toBeDefined();
+      expect(result.locName).toBeDefined();
+      return Promise.resolve();
+    })
+);
 
- const obj = {
- Data: new Buffer(helpers.generateRandomString()),
- Links: []
- };
- const ipfs_stack = config.nodes.map(node => ipfsAPI(node));
+test('update loc', () => {
 
- Promise.all(
- _.chain(ipfs_stack)
- .map(ipfs =>
- ipfs.object.put(obj)
- )
- .value()
- )
- .then((data) => {
- factory.Loc.old_hash = factory.Loc.hash;
- factory.Loc.hash = helpers.bytes32fromBase58(data[0].toJSON().multihash);
- return contracts.mint.setLOC(
- factory.Loc.name,
- factory.Loc.name,
- factory.Loc.website,
- factory.Loc.issueLimit,
- factory.Loc.hash,
- factory.Loc.expDate
- )
- })
- .then(result => {
- expect(result).toBeDefined();
- expect(result.tx).toBeDefined();
- return Promise.resolve();
- })
- });
+  const obj = {
+    Data: new Buffer(helpers.generateRandomString()),
+    Links: []
+  };
+  const ipfs_stack = config.nodes.map(node => ipfsAPI(node));
 
- test('fetch changes for loc via getLoc', () =>
- new Promise(res => {
- contracts.mint.allEvents({fromBlock: 0}).watch((err, result) => {
- if (result && result.event === 'UpdLOCValue') {
- expect(result.args.locName).toBeDefined();
- contracts.mint.getLOCByName(result.args.locName)
- .then(data => {
- if (data[4] === factory.Loc.hash) {
- res();
- }
- })
- }
- });
- })
- );
+  Promise.all(
+    _.chain(ipfs_stack)
+      .map(ipfs =>
+        ipfs.object.put(obj)
+      )
+      .value()
+  )
+    .then((data) => {
+      factory.Loc.old_hash = factory.Loc.hash;
+      factory.Loc.hash = helpers.bytes32fromBase58(data[0].toJSON().multihash);
+      return contracts_instances.LOCManager.setLOC(
+        factory.Loc.name,
+        factory.Loc.name,
+        factory.Loc.website,
+        factory.Loc.issueLimit,
+        factory.Loc.hash,
+        factory.Loc.expDate
+      )
+    })
+    .then(result => {
+      expect(result).toBeDefined();
+      expect(result.tx).toBeDefined();
+      return Promise.resolve();
+    })
+});
 
- test('fetch changes for loc via HashUpdate event', () =>
- new Promise(res => {
- contracts.mint.allEvents({fromBlock: 0}).watch((err, result) => {
- if (result && result.event === 'HashUpdate' && result.args.newHash === factory.Loc.hash) {
- res();
- }
- });
- })
- );
 
- test('validate new hash in mongo', () =>
- Promise.delay(2000)
- .then(() =>
- mongoose.model('hashupdates', new mongoose.Schema({
- oldHash: {type: mongoose.Schema.Types.Mixed},
- newHash: {type: mongoose.Schema.Types.Mixed}
- }
- )).findOne({oldHash: factory.Loc.old_hash, newHash: factory.Loc.hash})
- )
- .then(result => {
- expect(result).toBeDefined();
- return Promise.resolve();
- })
- );
- */
+test('fetch changes for loc via HashUpdate event', () =>
+  new Promise(res => {
+    contracts.LOCManager.at(contracts_instances.MultiEventsHistory.address)
+    .allEvents({fromBlock: 0}).watch((err, result) => {
+      if (result && result.event === 'HashUpdate' && result.args.newHash === factory.Loc.hash) {
+        res();
+      }
+    });
+  })
+);
+
+
+test('validate new hash in mongo', () =>
+  Promise.delay(2000)
+    .then(() =>
+      mongoose.model('hashupdates', new mongoose.Schema({
+          oldHash: {type: mongoose.Schema.Types.Mixed},
+          newHash: {type: mongoose.Schema.Types.Mixed}
+        }
+      )).findOne({oldHash: factory.Loc.old_hash, newHash: factory.Loc.hash})
+    )
+    .then(result => {
+      expect(result).toBeDefined();
+      return Promise.resolve();
+    })
+);
+
+
